Remove local TTS/STT engine, simplify to ElevenLabs-only
Drop Kokoro/mlx-audio TTS and Whisper/sherpa-onnx STT backends along
with their setup scripts, Python subprocess, type declarations, and
mock server. Simplify provider abstraction to ElevenLabs-only, move
TtsPlayer/SttProcessor interfaces to types.ts, inline bufferSentences
and writePcm into tts-elevenlabs.ts, update dashboard UI and routes,
and remove sherpa-onnx-node dependency.
