+++ README.md
@@ -1,6 +1,6 @@
 # Voice CC
 
-A Claude Code plugin for hands-free voice interaction with local speech-to-text, text-to-speech, and voice activity detection.
+A Claude Code plugin for hands-free voice interaction with ElevenLabs speech-to-text, text-to-speech, and voice activity detection.
 
 ## Project Structure
 
@@ -12,7 +12,7 @@ server/             Backend: voice pipeline + orchestration services
 dashboard/          Web UI (Vite + React) + API routes (Hono)
 lander/             Static landing page
 init/               Default prompt templates for new agents
-scripts/            Setup utilities (postinstall, local STT/TTS setup)
+scripts/            Setup utilities (postinstall)
 bin/                CLI entry point (voicecc command)
 ```
 
@@ -20,17 +20,16 @@ bin/                CLI entry point (voicecc command)
 
 ### Prerequisites
 
-- macOS with Apple Silicon (M1/M2/M3/M4)
+- macOS or Linux
 - Node.js 18+
-- Python 3.10+
-- Homebrew
+- An ElevenLabs API key
 
 ### Install
 
 ```bash
-# 1. Install system dependencies
+# 1. Install system dependencies (macOS)
 xcode-select --install
-brew install espeak-ng cloudflared
+brew install cloudflared
 
 # 2. Install Voice CC
 npm install -g voicecc
@@ -39,17 +38,13 @@ npm install -g voicecc
 voicecc
 ```
 
-The postinstall script handles sox, the Whisper model, Python venv, and TTS dependencies automatically.
-
 ## How It Works
 
-The voice loop runs locally with zero external API calls except to Claude:
-
-1. **Mic capture**: VPIO (macOS Voice Processing IO) records 16kHz mono PCM with echo cancellation
+1. **Mic capture**: VPIO (macOS) or PulseAudio (Linux) records 16kHz mono PCM with echo cancellation
 2. **Voice activity detection**: Silero VAD v5 detects speech segments
-3. **Speech-to-text**: sherpa-onnx (Whisper ONNX model) transcribes audio locally
+3. **Speech-to-text**: ElevenLabs Scribe API transcribes audio
 4. **Endpointing**: VAD silence-based turn detection
 5. **Claude inference**: Transcript sent to Claude Code Agent SDK session with streaming response
 6. **Narration**: Claude's response stripped of markdown and split into sentences
-7. **Text-to-speech**: Kokoro-82M via mlx-audio on Apple Silicon GPU (~8x realtime)
-8. **Speaker playback**: Audio output through VPIO at 24kHz with echo cancellation
+7. **Text-to-speech**: ElevenLabs streaming TTS API generates audio
+8. **Speaker playback**: Audio output through VPIO/PulseAudio at 24kHz
