+++ dashboard/src/components/VoiceProvidersPanel.tsx
@@ -2,14 +2,12 @@
  * Voice provider selection and status panel.
  *
  * Allows users to:
- * - Select TTS and STT providers via radio buttons
- * - See provider readiness status (Ready / Not Installed / Missing API Key / Unsupported Platform)
- * - Trigger on-demand setup for local providers (modal with live log output)
+ * - See provider readiness status (Ready / Missing API Key)
  * - Configure ElevenLabs API key and model settings (modal)
  * - Save all provider settings to .env
  */
 
-import { useState, useEffect, useCallback, useRef } from "react";
+import { useState, useEffect, useCallback } from "react";
 import { get, post } from "../api";
 
 // ============================================================================
@@ -20,16 +18,13 @@ const DEFAULT_VOICE_ID = "JBFqnCBsd6RMkjVDRZzb";
 const DEFAULT_MODEL_ID = "eleven_turbo_v2_5";
 const DEFAULT_STT_MODEL_ID = "scribe_v1";
 
-/** How often to poll for setup log updates (ms) */
-const SETUP_POLL_INTERVAL_MS = 1500;
-
 // ============================================================================
 // TYPES
 // ============================================================================
 
 interface ProviderStatus {
   ready: boolean;
-  reason?: "not_installed" | "missing_api_key" | "unsupported_platform";
+  reason?: "missing_api_key";
   detail?: string;
 }
 
@@ -45,15 +40,8 @@ interface ProvidersResponse {
   active: string;
 }
 
-interface SetupJobStatus {
-  running: boolean;
-  exitCode: number | null;
-  log: string;
-}
-
 type ModalState =
   | null
-  | { kind: "setup"; target: "local-tts" | "local-stt" }
   | { kind: "elevenlabs-tts" }
   | { kind: "elevenlabs-stt" };
 
@@ -90,18 +78,10 @@ function StatusBadge({ status }: { status: ProviderStatus }) {
     label = "Ready";
     bgColor = "rgba(34, 197, 94, 0.15)";
     textColor = "#22c55e";
-  } else if (status.reason === "not_installed") {
-    label = "Not Installed";
-    bgColor = "rgba(234, 179, 8, 0.15)";
-    textColor = "#eab308";
   } else if (status.reason === "missing_api_key") {
     label = "Missing API Key";
     bgColor = "rgba(239, 68, 68, 0.15)";
     textColor = "#ef4444";
-  } else if (status.reason === "unsupported_platform") {
-    label = "Unsupported Platform";
-    bgColor = "rgba(156, 163, 175, 0.15)";
-    textColor = "#9ca3af";
   } else {
     label = "Unknown";
     bgColor = "rgba(156, 163, 175, 0.15)";
@@ -197,131 +177,6 @@ function ProviderRow({
           {actionLabel}
         </button>
       )}
-    </div>
-  );
-}
-
-/** Modal for local provider setup with live log output */
-function SetupModal({
-  target,
-  onClose,
-  onComplete,
-}: {
-  target: "local-tts" | "local-stt";
-  onClose: () => void;
-  onComplete: () => void;
-}) {
-  const [log, setLog] = useState("");
-  const [running, setRunning] = useState(true);
-  const [exitCode, setExitCode] = useState<number | null>(null);
-  const preRef = useRef<HTMLPreElement>(null);
-  const startedRef = useRef(false);
-
-  // Auto-scroll to bottom when log updates
-  useEffect(() => {
-    if (preRef.current) {
-      preRef.current.scrollTop = preRef.current.scrollHeight;
-    }
-  }, [log]);
-
-  // Start the setup job and poll for updates
-  useEffect(() => {
-    if (startedRef.current) return;
-    startedRef.current = true;
-
-    let cancelled = false;
-
-    const run = async () => {
-      try {
-        const { jobId } = await post<{ jobId: string }>(
-          `/api/providers/setup/${target}`
-        );
-
-        // Small initial delay to let the process start
-        await new Promise((r) => setTimeout(r, 500));
-
-        const poll = async (): Promise<void> => {
-          if (cancelled) return;
-          try {
-            const status = await get<SetupJobStatus>(
-              `/api/providers/setup/status/${jobId}`
-            );
-            setLog(status.log);
-
-            if (status.running) {
-              await new Promise((r) => setTimeout(r, SETUP_POLL_INTERVAL_MS));
-              return poll();
-            }
-
-            setRunning(false);
-            setExitCode(status.exitCode);
-            if (status.exitCode === 0) onComplete();
-          } catch {
-            if (!cancelled) setRunning(false);
-          }
-        };
-
-        await poll();
-      } catch {
-        if (!cancelled) setRunning(false);
-      }
-    };
-
-    run();
-    return () => { cancelled = true; };
-  }, [target, onComplete]);
-
-  const title = target === "local-tts" ? "Setting up Local TTS" : "Setting up Local STT";
-
-  return (
-    <div className="modal-overlay visible" onClick={(e) => { if (e.target === e.currentTarget && !running) onClose(); }}>
-      <div className="modal" style={{ width: 640 }}>
-        <button className="modal-close" onClick={onClose} disabled={running}>&times;</button>
-        <h2>{title}</h2>
-        <p style={{ fontSize: 12, color: "var(--text-secondary)", marginBottom: 12 }}>
-          {running
-            ? "Installing dependencies. This may take a few minutes..."
-            : exitCode === 0
-              ? "Setup completed successfully."
-              : "Setup failed. Check the log below for details."}
-        </p>
-        <pre
-          ref={preRef}
-          style={{
-            background: "#1a1a2e",
-            color: "#e0e0e0",
-            padding: "12px 14px",
-            borderRadius: "4px",
-            fontSize: "11px",
-            lineHeight: "1.5",
-            maxHeight: "400px",
-            overflowY: "auto",
-            whiteSpace: "pre-wrap",
-            wordBreak: "break-word",
-            margin: 0,
-            border: "1px solid var(--border-color)",
-          }}
-        >
-          {log || "Starting setup...\n"}
-        </pre>
-        <div style={{ display: "flex", justifyContent: "flex-end", marginTop: 16 }}>
-          <button
-            onClick={onClose}
-            disabled={running}
-            style={{
-              padding: "6px 16px",
-              fontSize: "13px",
-              background: running ? "var(--bg-main)" : "var(--btn-primary-bg)",
-              color: running ? "var(--text-secondary)" : "var(--btn-primary-text)",
-              border: "1px solid var(--border-color)",
-              borderRadius: "4px",
-              cursor: running ? "default" : "pointer",
-            }}
-          >
-            {running ? "Running..." : "Close"}
-          </button>
-        </div>
-      </div>
     </div>
   );
 }
@@ -494,8 +349,8 @@ function ElevenLabsSttModal({
 export function VoiceProvidersPanel() {
   const [ttsProviders, setTtsProviders] = useState<ProviderInfo[]>([]);
   const [sttProviders, setSttProviders] = useState<ProviderInfo[]>([]);
-  const [activeTts, setActiveTts] = useState("local");
-  const [activeStt, setActiveStt] = useState("local");
+  const [activeTts, setActiveTts] = useState("elevenlabs");
+  const [activeStt, setActiveStt] = useState("elevenlabs");
   const [apiKey, setApiKey] = useState("");
   const [voiceId, setVoiceId] = useState(DEFAULT_VOICE_ID);
   const [modelId, setModelId] = useState(DEFAULT_MODEL_ID);
@@ -537,17 +392,6 @@ export function VoiceProvidersPanel() {
       .catch(() => {});
   }, []);
 
-  /** Refresh provider lists after a setup completes */
-  const refreshProviders = useCallback(async () => {
-    try {
-      const [tts, stt] = await Promise.all([fetchTtsProviders(), fetchSttProviders()]);
-      setTtsProviders(tts.providers);
-      setSttProviders(stt.providers);
-    } catch {
-      // Ignore refresh errors
-    }
-  }, []);
-
   /** Select a TTS provider and save immediately */
   const selectTts = useCallback((type: string) => {
     setActiveTts(type);
@@ -561,22 +405,12 @@ export function VoiceProvidersPanel() {
   }, [saveSettings]);
 
   /** Get action button config for a provider row */
-  const getAction = (providerType: string, status: ProviderStatus, section: "tts" | "stt") => {
-    if (providerType === "local" && !status.ready && status.reason === "not_installed") {
-      const target = section === "tts" ? "local-tts" : "local-stt";
-      return {
-        label: "Setup",
-        onAction: () => setModal({ kind: "setup" as const, target: target as "local-tts" | "local-stt" }),
-      };
-    }
-    if (providerType === "elevenlabs") {
-      const kind = section === "tts" ? "elevenlabs-tts" : "elevenlabs-stt";
-      return {
-        label: "Configure",
-        onAction: () => setModal({ kind: kind as "elevenlabs-tts" | "elevenlabs-stt" }),
-      };
-    }
-    return {};
+  const getAction = (_providerType: string, _status: ProviderStatus, section: "tts" | "stt") => {
+    const kind = section === "tts" ? "elevenlabs-tts" : "elevenlabs-stt";
+    return {
+      label: "Configure",
+      onAction: () => setModal({ kind: kind as "elevenlabs-tts" | "elevenlabs-stt" }),
+    };
   };
 
   return (
@@ -643,15 +477,6 @@ export function VoiceProvidersPanel() {
           })}
         </div>
       </div>
-
-      {/* Setup Modal (local provider installation) */}
-      {modal?.kind === "setup" && (
-        <SetupModal
-          target={modal.target}
-          onClose={() => setModal(null)}
-          onComplete={refreshProviders}
-        />
-      )}
 
       {/* ElevenLabs TTS Configuration Modal */}
       {modal?.kind === "elevenlabs-tts" && (
