+++ server/voice/browser-server.ts
@@ -18,8 +18,6 @@
 import "dotenv/config";
 
 import { createServer, request as httpRequest } from "http";
-import { homedir } from "os";
-import { join } from "path";
 
 import { WebSocketServer } from "ws";
 
@@ -31,7 +29,7 @@ import type { IncomingMessage, ServerResponse } from "http";
 import type { Duplex } from "stream";
 import type { WebSocket } from "ws";
 import type { VoiceSession } from "./voice-session.js";
-import type { TtsProviderConfig, SttProviderConfig, TtsProviderType, SttProviderType } from "./types.js";
+import type { TtsProviderConfig, SttProviderConfig } from "./types.js";
 
 // ============================================================================
 // CONSTANTS
@@ -46,9 +44,7 @@ const BROWSER_INTERRUPTION_THRESHOLD_MS = 1500;
 /** Ping interval to keep WebSocket connections alive through tunnel (ms) */
 const PING_INTERVAL_MS = 30_000;
 
-/** Read provider selection and ElevenLabs config from environment */
-const TTS_PROVIDER = (process.env.TTS_PROVIDER ?? "local") as TtsProviderType;
-const STT_PROVIDER = (process.env.STT_PROVIDER ?? "local") as SttProviderType;
+/** Read ElevenLabs config from environment */
 const ELEVENLABS_API_KEY = process.env.ELEVENLABS_API_KEY ?? "";
 const ELEVENLABS_VOICE_ID = process.env.ELEVENLABS_VOICE_ID ?? "WrjxnKxK0m1uiaH0uteU";
 const ELEVENLABS_MODEL_ID = process.env.ELEVENLABS_MODEL_ID ?? "eleven_turbo_v2_5";
@@ -56,15 +52,13 @@ const ELEVENLABS_STT_MODEL_ID = process.env.ELEVENLABS_STT_MODEL_ID ?? "scribe_v
 
 /** TTS provider configuration built from env vars */
 const ttsProvider: TtsProviderConfig = {
-  provider: TTS_PROVIDER,
-  local: { model: "prince-canuma/Kokoro-82M", voice: "af_heart" },
+  provider: "elevenlabs",
   elevenlabs: { apiKey: ELEVENLABS_API_KEY, voiceId: ELEVENLABS_VOICE_ID, modelId: ELEVENLABS_MODEL_ID },
 };
 
 /** STT provider configuration built from env vars */
 const sttProvider: SttProviderConfig = {
-  provider: STT_PROVIDER,
-  local: { modelPath: join(homedir(), ".claude-voice-models", "whisper-small") },
+  provider: "elevenlabs",
   elevenlabs: { apiKey: ELEVENLABS_API_KEY, modelId: ELEVENLABS_STT_MODEL_ID },
 };
