--- server/voice/sherpa-onnx-node.d.ts
@@ -1,4 +1,0 @@
-declare module "sherpa-onnx-node" {
-  const sherpa: any;
-  export default sherpa;
-}
