+++ server/voice/stt-provider.ts
@@ -1,39 +1,19 @@
 /**
  * STT provider factory and readiness checks.
  *
- * Routes STT creation to the correct provider implementation based on config.
- * Checks provider readiness (model files, API keys) for dashboard status.
+ * Routes STT creation to the ElevenLabs provider implementation.
+ * Checks provider readiness (API keys) for dashboard status.
  *
  * Responsibilities:
- * - Create an SttProcessor for the configured provider (local or ElevenLabs)
- * - Check provider readiness (model files exist, API keys set)
+ * - Create an SttProcessor for the configured provider
+ * - Check provider readiness (API keys set)
  * - Provide static metadata about available STT providers
  */
 
-import { existsSync } from "fs";
-import { homedir } from "os";
-import { join } from "path";
-
-import { createLocalStt } from "./stt.js";
 import { createElevenlabsStt } from "./stt-elevenlabs.js";
 import { readEnv } from "../services/env.js";
 
-import type { SttProcessor } from "./stt.js";
-import type { SttProviderType, SttProviderConfig, ProviderStatus } from "./types.js";
-
-// ============================================================================
-// CONSTANTS
-// ============================================================================
-
-/** Standard path where local Whisper model files are stored */
-const LOCAL_STT_MODEL_DIR = join(homedir(), ".claude-voice-models", "whisper-small");
-
-/** Required model files for local Whisper STT */
-const REQUIRED_MODEL_FILES = [
-  "small.en-encoder.int8.onnx",
-  "small.en-decoder.int8.onnx",
-  "small.en-tokens.txt",
-];
+import type { SttProcessor, SttProviderType, SttProviderConfig, ProviderStatus } from "./types.js";
 
 // ============================================================================
 // INTERFACES
@@ -49,8 +29,6 @@ export interface SttProviderInfo {
   name: string;
   /** Short description of the provider */
   description: string;
-  /** Platform required for this provider (undefined = any platform) */
-  requiresPlatform?: "darwin";
   /** Environment variable name for the API key (undefined = no key needed) */
   requiresApiKey?: string;
 }
@@ -69,7 +47,6 @@ export interface CreateSttOptions {
 
 /**
  * Create an SttProcessor for the configured provider.
- * Routes to the local Whisper provider or ElevenLabs cloud provider.
  *
  * @param options - Provider config with per-provider settings
  * @returns An SttProcessor instance ready for transcription
@@ -79,9 +56,6 @@ export async function createSttForProvider(options: CreateSttOptions): Promise<S
   const { providerConfig } = options;
 
   switch (providerConfig.provider) {
-    case "local":
-      return createLocalStt(providerConfig.local.modelPath);
-
     case "elevenlabs":
       return createElevenlabsStt({
         apiKey: providerConfig.elevenlabs.apiKey,
@@ -95,30 +69,13 @@ export async function createSttForProvider(options: CreateSttOptions): Promise<S
 
 /**
  * Check whether an STT provider is ready to use.
- *
- * Local: checks that the 3 required Whisper model files exist at the standard path.
- * ElevenLabs: checks ELEVENLABS_API_KEY is set in .env.
+ * Checks ELEVENLABS_API_KEY is set in .env.
  *
  * @param providerType - The provider to check
  * @returns Readiness status with reason if not ready
  */
 export async function getSttProviderStatus(providerType: SttProviderType): Promise<ProviderStatus> {
   switch (providerType) {
-    case "local": {
-      const missingFiles = REQUIRED_MODEL_FILES.filter(
-        (file) => !existsSync(join(LOCAL_STT_MODEL_DIR, file))
-      );
-
-      if (missingFiles.length > 0) {
-        return {
-          ready: false,
-          reason: "not_installed",
-          detail: `Missing model files in ${LOCAL_STT_MODEL_DIR}: ${missingFiles.join(", ")}`,
-        };
-      }
-      return { ready: true };
-    }
-
     case "elevenlabs": {
       const env = await readEnv();
       if (!env.ELEVENLABS_API_KEY) {
@@ -139,11 +96,6 @@ export async function getSttProviderStatus(providerType: SttProviderType): Promi
  */
 export function getAvailableSttProviders(): SttProviderInfo[] {
   return [
-    {
-      type: "local",
-      name: "Local Whisper",
-      description: "On-device STT via sherpa-onnx Whisper ONNX model (offline batch mode)",
-    },
     {
       type: "elevenlabs",
       name: "ElevenLabs Scribe",
